+++ src/components/store/StoreDetail.jsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import { useParams } from 'react-router-dom';
-import { useQuery } from '@tanstack/react-query';
-import { storeQueryKey } from '../../constants/index';
-import { fetchStore } from '../../api/stores';
-import { Title, Votes, DetailSide } from './index';
+import { useStore } from '../../hooks/index';
+import { Header, Votes, DetailSide } from './index';
 
 const StoreDetailContainer = styled.div`
   width: 100%;
@@ -51,11 +49,9 @@ const VoteCntMsg = styled.span`
   }
 `;
 
-const storeQuery = storeid => ({ queryKey: [...storeQueryKey, storeid], queryFn: fetchStore(storeid) });
-
-const Details = ({ archivedCntState, setArchiveCntState, addBookMark, deleteBookMark }) => {
+const StoreDetail = ({ archivedCntState, setArchiveCntState, addBookMark, deleteBookMark }) => {
   const { id } = useParams();
-  const { data: storeData } = useQuery(storeQuery(id));
+  const { data: storeData } = useStore(id);
 
   React.useEffect(() => {
     setArchiveCntState(storeData?.archivesCount);
@@ -63,12 +59,11 @@ const Details = ({ archivedCntState, setArchiveCntState, addBookMark, deleteBook
 
   return (
     <StoreDetailContainer className="storedetail">
-      <Title storeData={storeData} addBookMark={addBookMark} deleteBookMark={deleteBookMark} />
+      <Header storeData={storeData} addBookMark={addBookMark} deleteBookMark={deleteBookMark} />
       <SubTitle>
         <FirstVoteUser>
           최초 투표자 : <UserName>{storeData.firstVoteUser}</UserName>
         </FirstVoteUser>
-
         <VoteCntMsg>
           투표 <span>{storeData.totalVotesCnt}</span>개
         </VoteCntMsg>
@@ -82,4 +77,4 @@ const Details = ({ archivedCntState, setArchiveCntState, addBookMark, deleteBook
   );
 };
 
-export default Details;
+export default StoreDetail;
