+++ src/components/store/index.js
@@ -1,10 +1,10 @@
 export { default as StorePositionMap } from './StorePositionMap';
-
-export { default as Title } from './Title';
+export { default as Header } from './Header';
 export { default as DetailSide } from './DetailSide';
-export { default as Details } from './Details';
+export { default as StoreDetail } from './StoreDetail';
 export { default as Votes } from './Votes';
-export { default as Comments } from './Comments';
+export { default as CommentsList } from './CommentsList';
 export { default as CommentsTextArea } from './CommentsTextArea';
-export { default as CommentsData } from './CommentsData';
+export { default as Comment } from './Comment';
+export { default as Pagination } from './Pagination';
 export { default as StoreImgUpload } from './StoreImgUpload';
